requirements.txt
grandalf
langchain-exa
langchain-google-genai>=0.1.0
langchain-openai
langgraph
pydantic-ai
thirdweb-ai[all]>=0.1.2
python-dotenv
langchain_core
rich 
black 
jupyterlab 
jupyterlab_widgets
fastapi>=0.105.0
uvicorn>=0.24.0
pydantic>=2.0.0
email-validator
motor>=3.3.0
pymongo>=4.5.0
dnspython>=2.4.2